requirements.txt
packaging==24.2
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.1/libexec/wheel-0.45.1-py3-none-any.whl#sha256=da46333d5dcbde6e20cf7e2f8fff9e9ce76e8c94dc4afd6fb95fc4bc2745fb5e
moviepy
openai-whisper
ffmpeg-python==0.2.0
pysrt==1.1.2